Add support for assignment ("walrus") expressions
These are fairly simple, but the implementation intentionally chooses to skip sanitizing values because there is no "out" like there is with
assignment to preserve constexprs using an annotation. I would argue
that sanitization is actually undesirable in general (see #8749 for
rationale) but that is a discussion for another day.